+++ src/core/configuration.c
@@ -72,9 +72,9 @@ MsQuicConfigurationOpen(
     if (Configuration == NULL) {
         QuicTraceEvent(
             AllocFailure,
-            "Allocation of '%s' failed. (%zu bytes)",
+            "Allocation of '%s' failed. (%llu bytes)",
             "QUIC_CONFIGURATION" ,
-            sizeof(QUIC_CONFIGURATION));
+            (uint64_t)sizeof(QUIC_CONFIGURATION));
         Status = QUIC_STATUS_OUT_OF_MEMORY;
         goto Error;
     }
