+++ src/core/library.c
@@ -294,8 +294,8 @@ MsQuicLibraryInitialize(
     if (MsQuicLib.DefaultCompatibilityList == NULL) {
         QuicTraceEvent(
             AllocFailure,
-            "Allocation of '%s' failed. (%" PRIu32 " bytes)", "default compatibility list",
-            CompatibilityListByteLength);
+            "Allocation of '%s' failed. (%llu bytes)", "default compatibility list",
+            (uint64_t)CompatibilityListByteLength);
         Status = QUIC_STATUS_OUT_OF_MEMORY;
         goto Error;
     }
@@ -338,8 +338,8 @@ MsQuicLibraryInitialize(
     if (MsQuicLib.PerProc == NULL) {
         QuicTraceEvent(
             AllocFailure,
-            "Allocation of '%s' failed. (%zu bytes)", "connection pools",
-            MsQuicLib.ProcessorCount * sizeof(QUIC_LIBRARY_PP));
+            "Allocation of '%s' failed. (%llu bytes)", "connection pools",
+            (uint64_t)(MsQuicLib.ProcessorCount * sizeof(QUIC_LIBRARY_PP)));
         Status = QUIC_STATUS_OUT_OF_MEMORY;
         goto Error;
     }
@@ -981,9 +981,9 @@ QuicLibrarySetGlobalParam(
         if (NewConfig == NULL) {
             QuicTraceEvent(
                 AllocFailure,
-                "Allocation of '%s' failed. (%" PRIu32 " bytes)",
+                "Allocation of '%s' failed. (%llu bytes)",
                 "Execution config",
-                BufferLength);
+                (uint64_t)BufferLength);
             Status = QUIC_STATUS_OUT_OF_MEMORY;
             break;
         }
