+++ src/core/registration.c
@@ -65,9 +65,9 @@ MsQuicRegistrationOpen(
     if (Registration == NULL) {
         QuicTraceEvent(
             AllocFailure,
-            "Allocation of '%s' failed. (%zu bytes)",
+            "Allocation of '%s' failed. (%llu bytes)",
             "registration",
-            sizeof(QUIC_REGISTRATION) + AppNameLength + 1);
+            (uint64_t)(sizeof(QUIC_REGISTRATION) + AppNameLength + 1));
         Status = QUIC_STATUS_OUT_OF_MEMORY;
         goto Error;
     }
